Moved some convenience functions to their own file
