+++ src/syntax/cron/SyntaxNode.ts
@@ -1,8 +1,6 @@
 /**
  * A node in the cron syntax tree.
  */
-import { DualExpression, StringLiteral } from "@/syntax";
-
 export default interface SyntaxNode {
   /**
    * Get the generated value for this node
