+++ src/syntax/index.ts
@@ -1,2 +1,3 @@
 export * from "./cron";
 export * from "./quartz";
+export * from "./common";
